supposedly added sse event tracking for game and club score updates; added testing.md file to document different test mechanisms

+++ services/espn.go
@@ -52,7 +52,7 @@ type ESPNStatus struct {
 	Type         ESPNStatusType `json:"type"`
 	Period       int            `json:"period"`
 	DisplayClock string         `json:"displayClock,omitempty"`
-	Clock        int            `json:"clock,omitempty"`
+	Clock        float64        `json:"clock,omitempty"` // Changed to float64 to handle ESPN's 0.0 values
 }
 
 type ESPNStatusType struct {
